renames worker files for easier debugging
