+++ src/DataChart/DataChartWorker.ts
@@ -1,4 +1,4 @@
 import { WorkerService } from "../Webworkers/WorkerService";
-import * as DataChartService from "./service";
+import * as DataChartService from "./requests";
 
 new WorkerService("DataChartService", DataChartService);
