+++ src/EVCharging/EVWorker.ts
@@ -1,4 +1,4 @@
 import { WorkerService } from "../Webworkers/WorkerService";
-import * as EVChargingService from "./service";
+import * as EVChargingService from "./requests";
 
 new WorkerService("EVChargingService", EVChargingService);
