+++ src/Monitors/MonitorsWorker.ts
@@ -1,4 +1,4 @@
 import { WorkerService } from "../Webworkers/WorkerService";
-import * as MonitorsService from "./service";
+import * as MonitorsService from "./requests";
 
 new WorkerService("MonitorsService", MonitorsService);
